refactor: refactoring AutoUpdateInterceptor
