+++ src/WeihanLi.EntityFramework/Interceptors/AutoUpdateInterceptor.cs
@@ -1,17 +1,7 @@
-﻿using Microsoft.EntityFrameworkCore.ChangeTracking;
-using Microsoft.EntityFrameworkCore.Diagnostics;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading;
-using System.Threading.Tasks;
+﻿using Microsoft.EntityFrameworkCore.Diagnostics;
+using WeihanLi.EntityFramework.Services;
 
 namespace WeihanLi.EntityFramework.Interceptors;
-
-public interface IEntitySavingHandler
-{
-    void Handle(EntityEntry entityEntry);
-}
 
 public sealed class AutoUpdateInterceptor(IEnumerable<IEntitySavingHandler> handlers) : SaveChangesInterceptor
 {
@@ -24,7 +14,7 @@ public sealed class AutoUpdateInterceptor(IEnumerable<IEntitySavingHandler> hand
     }
 
     public override ValueTask<InterceptionResult<int>> SavingChangesAsync(DbContextEventData eventData, InterceptionResult<int> result,
-        CancellationToken cancellationToken = new CancellationToken())
+        CancellationToken cancellationToken = default)
     {
         OnSavingChanges(eventData);
         return base.SavingChangesAsync(eventData, result, cancellationToken);
