+++ src/WeihanLi.EntityFramework/Services/IEntitySavingHandler.cs
@@ -1,0 +1,9 @@
+﻿using Microsoft.EntityFrameworkCore.ChangeTracking;
+
+namespace WeihanLi.EntityFramework.Services;
+
+public interface IEntitySavingHandler
+{
+    void Handle(EntityEntry entityEntry);
+}
+
